Rewrite config chain loading to build chain recursively to keep caching readable.
